これからプラグインを分けるにあたり、共通で使用するEnumとStructはDefinedヘッダーへ移行。他プラグイン使用時に同名で起動できない問題が発生したため、一部Enumの名前変更
