+++ Sample/Plugins/VoicevoxEngine/Source/VoicevoxEngine/Private/VoicevoxCoreUtil.cpp
@@ -16,14 +16,6 @@
 
 DEFINE_LOG_CATEGORY(LogVoicevoxEngine);
 
-void FVoicevoxAudioQuery::AddStructReferencedObjects(FReferenceCollector& Collector) const
-{
-	if (const UScriptStruct* Struct = StaticStruct(); Struct != nullptr && Pointer != nullptr)
-	{
-		Collector.AddReferencedObjects(Struct, Pointer);
-	}
-}
-
 /**
  * @brief VOICEVOX CORE 初期化
  */
