+++ Sample/Plugins/VoicevoxEngine/Source/VoicevoxEngineEditor/Public/VoicevoxQueryTypeActions.h
@@ -7,6 +7,7 @@
  */
 
 #pragma once
+
 #include "AssetTypeActions_Base.h"
 
 class FVoicevoxQueryTypeActions : public FAssetTypeActions_Base
